requirements.txt
google-generai==1.19.0
pillow==10.4.0
pytest==8.3.4